CICD-13 Add push all pranches (except main/develop) to dev repo

+++ src/localbase.py
@@ -59,8 +59,11 @@ class LocalBaseWorker:
         return local_name
 
     def remove_database(self, local_name):
-        self.db_session.query(BasesTable).filter_by(local_name=local_name).delete()
-        self.db_session.commit()
+        # Don't execute in local database when local_name is False
+        # In Database/Put except block we pass none
+        if local_name:
+            self.db_session.query(BasesTable).filter_by(local_name=local_name).delete()
+            self.db_session.commit()
 
     def get_database_and_connection(self, local_base_name: str):
         database_obj = (
@@ -185,15 +188,16 @@ class LocalBaseWorker:
         ]
 
     def get_admin_tokens_objects_list(self) -> List[TokenTable]:
-        # Returns only tokens with access
+        # Returns only tokens with admin access
         return [
             i for i in get_existing_data(self.db_session, TokenTable) if i.admin_access
         ]
 
     def get_tokens_list(self) -> List[str]:
+        # Returns all tokens (with admin access and without admin access)
         return [i.token for i in get_existing_data(self.db_session, TokenTable)]
 
-    def get_tokens_names_list(self) -> List[str]:
+    def get_user_tokens_names_list(self) -> List[str]:
         return [i.name for i in self.get_user_tokens_objects_list()]
 
     def add_table_for_token(
@@ -205,6 +209,15 @@ class LocalBaseWorker:
         local_table_name: str = None,
         token_name: str = None,
     ):
+        """
+        Add token access for select requests for table
+        :param token: token to grant access
+        :param database_name: local database name where table located
+        :param folder_name: folder name in database where table located
+        :param table_name: table name in database where table located
+        :param local_table_name: local name of table (if pass local_table_name than database_name/folder_name/table_name will be ignored) # noqa: E501
+        :param token_name: token name to grant access (if passed token - token_name will be ignored)
+        """
         if not local_table_name:
             local_table_name = self.get_local_table_name(
                 database_name=database_name,
@@ -231,6 +244,10 @@ class LocalBaseWorker:
         folder_name: str,
         table_name: str,
     ):
+        # In mysql folder has same name as database
+        if self.get_database_object(database_name).type == "mysql":
+            folder_name = database_name
+
         table_object = (
             self.db_session.query(TablesInfoTable)
             .filter_by(
